% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_to_colls.R
\name{datetime_to_cols}
\alias{datetime_to_cols}
\title{DateTime to columns}
\usage{
datetime_to_cols(datetime, timeOfDay = "danish")
}
\arguments{
\item{datetime}{A datetime object}
}
\value{
A Data Frame including the following columns:

\itemize{
  \item \strong{datetime:} The inputtet datetime
}
}
\description{
Extract various info from DateTime such as weekday, time of day, during working hours, etc.
}
